tcenv/magazine/requirements.txt
asgiref==3.8.1
Django==5.0.7
django-rest-framework==0.1.0
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.15.2
dnspython==2.6.1
drf-yasg==1.21.7
inflection==0.5.1
mongoengine==0.28.2
packaging==24.1
pymongo==4.8.0
pytz==2024.1
PyYAML==6.0.1
setuptools==70.3.0
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
